Begin to seperate out dom maniplation files
